High: PE: Bug LF185 - Prevent pointless master/slave instance shuffling by ignoring the master-pref of stopped instances After consultation with lmb, stopped master instances should have no master-pref value. In the absense of a rsc_location constraint, this will prevent them from being considered for promotion until they are started again.
This resolves Bug LF1852 in which a previous master (about to be started) had
a higher master-pref (causing it to be allocated first).  However when it came
time to allocate it, it preferded the location of the current master (due to
colocation constraints) which resulted in pointless shuffling.

Mercurial revision: 5f1a9b74d493aeaf27c2f27682b83525c2b10550