High: ais: Dampen membership changes when we loose more than half the known nodes Starting a new node can occasionally cause the openais membership size to hiccup.  Eg.   5 -> 1 -> 6 These hiccups usually last between 1 and 2 seconds but can seriously destabilize the cluster - particularly when STONITH is enabled (the "lost" nodes get shot).
To mask this, we add a 2s delay whenever the membership size drops by more than
half the previous size.  If, before that 2s has elapsed, a new membership is
delivered that doesn't meet this criteria - then it is delivered immediately.

We also delay the initial membership notification if it places us in a
cluster-of-one.

Mercurial revision: d80e066655018c4c04568c52b01022396e40d86f