High: crmd: Don't hide stop events that time out - allowing faster recovery in the presence of overloaded hosts
A long time ago in a galaxy far away, some messaging layers used to
loose quite a few actions, including stops.  About the same time, we
decided that fencing because a stop action was lost wasn't a good
idea.

The rationale was that if the operation eventually completed, it would
end up in the CIB anyway.  And even if it didn't, the PE would
continue to try the operation again until the whole node fell over at
which point it would get shot anyway.

Things have improved since then and so, in the interest of speeding
up recovery in these situations, it is time to stop treating stop
operations differently.

Mercurial revision: dd8e37df3e963e983fe3a0a3f27d3deb487e8dd0