High: Core: Provide significant performance improvements by implementing versioned diffs and digests
The versioning allows us to implement a new digest algorithm without breaking
backwards compatability.

We also attach the current version when generating diffs so that we know
which algorithm to use when validating the result against the supplied digest.

The sorting associated with v1 digest creation accounted for 23% of
the CIB's CPU usage on the server. The new algorithm drops this.

Mercurial revision: 38d89f3740ee94b65c346357a8d7b03aa255ce1c