High: crmd: Ensure the CIB is always writable on the DC by removing a timing hole
Somehow the slave call was processed after the master call which left the
CIB r/o until the next election.

There is no need for the DC to be changing the CIB state on its peers,
the election process has handled this for years.  So just remove the
call to slave_all

Mercurial revision: 945a67346ddda44b181003370dee5e56c0d3b128