Medium: Shell: node clearstate must flush LRM state too (bnc#653795)
Note that "clearstate" is different from delete; it leaves the node in
the configuration, or even puts a clean state in if it hasn't been seen
in that transition yet.

So we first modify (perhaps create) the node_state and then delete any
we might have remembered.

Mercurial revision: 5ce2879aa0d5f43d01629bc20edc6868a9352002