High: Core: Bug cl#5016 - Prevent failures in recurring ops from being lost
Failures in recurring ops were not being saved to the last_failure
entry and were lost once a stop occurred.

However doing this means there can be more than both a success and
failure for the same op, so the PE needed to be able to determin the
relative ordering.

We also had to bump the feature set so that we wouldn't send updates
that would confuse older PEs.

Lastly, we also have to avoid deleting the last_failure when cancelling
the recurring operation.