High: Core: Strip text nodes from on disk xml files
libxml2's parser treats the whitespace between:
  <tag>
  </tag>
as significant.

This affects the output which we use for calculating the CIB digests.
In some situations, this can result in the CIB needing a full sync
to all nodes after every update - and lots of CPU - even though the
content we care about is the same.

No part of the system uses text nodes, so they are safe to remove
unconditionally.