High: crmd: Remove rusage differences of less than a second from election calculations
Given nodes A, B, C and a dampening threshold T microseconds;

If T/2 < Br - Ar < T, and
   T/2 < Cr - Br < T
Where Ar == rusage of node A etc.

Then the election will never produce a winner, as C wins against A on
rusage but looses to B on node name (since the rusage difference is not
sufficiently significant).  B looses to A but wins against C (both on
name).

The only remedy is to use either the rusage outputs directly
(undampened), or drop it entirely as the higher the value of T, the more
likely the situation becomes

Microsecond resolutuion is too volitile, just counting the vote would
perterb its value enough to affect the next comparision. Happily we can
drop it and use only the value for tv_sec.