High: PE: Move the ticket state section to "/cib/status/tickets/ticket_state"
Move the ticket state section:
--- "/cib/status/tickets/instance_attributes"
+++ "/cib/status/tickets/ticket_state"

Pengine still recognizes the deprecated section. If there's any
attribute appears in the new tickets_state section, it'll take over
precedence over the one in the deprecated section. Actually, even
more, once crm_ticket is invoked to set an attribute, it'll clear the
one in the deprecated section.