High: PE: Correctly migrate a resource that has just migrated
After a migration, we run a 'stop' on the original node, mostly to
clean things up in the CIB.
If the transition was interrupted after the B.migrate_from but
before A.stop, then there will be two 'stop's for the resource and
get_first_named_action() might return the 'wrong' one - leading us
to think we need to migrate from the old host.

Ie.
  Migrate lustre (Started lustre01-left -> lustre03-left)
  Migrate lustre (Started lustre01-left -> lustre04-left)

instead of:

  Migrate lustre (Started lustre01-left -> lustre03-left)
  Migrate lustre (Started lustre03-left -> lustre04-left)