Low: Build: rhbz#805147 - Use default value for HB_DAEMON_DIR define when clusterglue does not provide one.
The clusterglue devel package in rhel 6 is removing support for the
HA_LIBHBDIR define in glue_config.h.  That define causes conflicts with
multilib support as it is different depending on what architecture is
being used.  Most commonly this define is used by pacemaker
as the location of the lrmd daemon and is stored in HB_DAEMON_DIR.  To
account for the fact HA_LIBHBDIR may not be present in glue_config.h,
pacemaker now defaults HB_DAEMON_DIR to either /usr/lib/heartbeat
or /usr/lib64/heartbeat when HA_LIBHBDIR is not found.