High: crmd: cl#5053 - Fixes fail-count not being updated properly.
When the failcount was being updated for a node, the update
sent to the attrd process contained the wrong uname for the
node.  This was a result of the get_uname() function not
being updated to handle the pcmk_cluster_corosync cluster type.
To resolve this, similar logic to the get_uuid() function (the
inverse of the get_uname()) function was set in place in get_uname()
to return the correct value.