Low: pengine: cl#5056 - Do not allow order dependencies between stonith and non-stonith resources.
There are internal pseudo actions that aggregate the completion
of stonith resources before other resources.  Any attempt to create
an order dependency between a stonith resouce and a non-stonith
resouce has the potential to generate an invalid transition graph.
Because of this, order dependencies between stonith and non-stonith
resouces are now ignored.