High: fencing: If multiple fencing devices are present, don't stop with the first one if it fails to fence the node.
When a remote stonith operation does not explicitly say what
device should be used to fence a node, a list of all the capable
devices is generated.  If the first device in that list failed
to fence the node, none of the other devices in the list would
be attempted.  This patch fixes that.