High: cib: Add the ability to use sockets and/or shared memory for IPC
When using sockets, since it is not possible in advance to know if a
send will block, if the server is interrupted while sending a message,
the caller will block until the send completes.  Otherwise the data sent
doesn't match the advertised amount and all hell breaks looks.

Likewise, the socket client must block when a partial message is recieved.

However, because shared memory knows how much space is available before
it starts sending, it can return with -EAGAIN /before/ sending anything.
Thus the partial message problem goes away.

For the crmd, we want the later behavior.