Low: crmd: cl#5064 - Loss of cpg in process list should be reflected in node status.
The node's ha status should be updated to 'dead' when cpg is lost.
When corosync is in use, it is possible that the membership layer
will not report a node has entered/left the cluster when a node
leaves the cpg.  This occurs for a node when pacemaker terminates
unexpectedly while corosynce remains up.  Because of this the
ha status remained "active" even though the cpg for that node
was known to be gone.