High: cib: cl#5026 - Synced cib updates should not return until the cpg broadcast is complete.
This fixes a race condition on shutdown where the final lrm
operation updates sent to the cib are not broadcast to all the
nodes before the crmd leaves the cpg.  This results in the final
lrm operations being ignored, which makes it look like the node
did not terminate properly.  When this occurs the node that just
shutdown gets fenced even though it didn't need to be. To fix this
local notifications from the cib are not sent out to the crmd
until the broadcast is complete.