High: pengine: cl#5069 - Honor 'on-fail=ignore' even when operation is disabled.
When a monitor operation's on-fail option is set to ignore, failures should
not result in recovery of the resource.  This works properly when the monitor
operation is enabled, but if the monitor operation fails and is then disabled,
the on-fail=ignore option is not honored. The result is the resource
gets recovered even though the monitor failure should be ignored. This issue
occurred because disabled operations were not being unpacked.  The
disabled monitor operation would be created, but instead of filling in the
action's attributes from the xml, defaults would be used instead.