High: crmd: cl#5068 - Fixes crm_node -R option so it works with corosync 2.0
The -R option for crm_node previously worked by connecting to the
cluster and sending out a command to all the nodes to clear
their peer cache for a specific node.  Now that cpg is used for
membership and communication between all the components across
the cluster, this method would not work.  Crm_node can not connect
to the crmd cpg to tell everyone to do something because it confuse
the membership layer of the crmd.  To enable this functionality
with use of corosync 2.0, a new ipc operation is created for crmd.
This operation allows crm_node to tell the local crmd process to
broadcast to everyone in the cgp to clear a node from their cache.