Low: crmd: Removes unnecessary error message from lrmd event processing.
It is safe to ignore cancelled operations that do not have a pending
operation associated with them in the lrm glue code.  With the previous
lrmd, we did not expect to hear about a cancelled operation after
executing the client command to cancel it.  The new lrmd will
notify everyone that the cancel completed (including the client
originally executing the cancel) through the ipc event channel.