Medium: PE: Simplify anonymous clone handling by discarding recorded instance numbers
When processing somechild:1 from the status section, drop the :1 and
use whatever clone instance we deem appropriate.

The logic for ensuring the commands are sent back to the node with
the name it expects (ie. somechild:1) already exists.

This is the first step towards removing instance numbers from the graph
for anonymous clones.