High: fencing: cl#5073 - Add 'off' as an valid value for stonith-action option.
The fencing agent api specifies that 'off' should be used
instead of 'poweroff' when defining new fencing devices.
The stonith-action option in the cib only accepts 'reboot' and
'poweroff'.  Since 'off' is the documented operation, this
patch adds 'off' as a valid stonith-action as well. The
'poweroff' operation is left as a valid value, but is documented
as existing only for legacy purposes now.  The default stonith-action
remains 'reboot'.