High: attrd: Differentiate between updating all our attributes and everybody updating all theirs too
When we first connect to the cib or our local cib was replaced, we
should only update our own entries.  If anyone else was affected they'll
notice and do the same.

Having every affected node trigger a site-wide refresh slams the cib and
produces so many events that client can't keep up.