High: fencing: Return cached dynamic target list for busy devices.
When determining if a device with a dynamic target list can fence
a node, a query is done to the device to return the list.  If that
device is busy, it is possible stonith will block indefinitely
waiting for that list.  Now instead of blocking, stonith uses
a cached target list when the devices can not be queried
instead of blocking.