High: ipc: Stop waiting for ipc reply on server disconnect.
If a client is disconnected between sending a request and
waiting for the reply, detect the disconnect and stop
waiting for the reply.  If for some reason a request causes
the server to crash, this scenario will definitely happen.
With out breaking from the reply wait loop on disconnect,
the recv loop will cause a flood of errors to the logfile for the
remaining duration of the timeout period.