High: fencing: cl#5092 - Always timeout stonith operations if timeout period expires.
When a stonith operation's timeout period is reached, stonithd sends
the child process SIGTERM and then SIGKILL if the process is still around.
Later on the WIFSIGNALED macro is used to determine if the process was
sent one of these two signals. The problem with that macro is that it will
only return the signal if the child process ignored it.  If a process detects
the SIGTERM signal and exists, WIFSIGNALED returns nothing so we don't
report the timeout occurred.

Now if the operation timer expires, stonith treats that operation as being
timed out regardless of what WIFSIGNALED returns.