Medium: Fencing: Allow node suicide when fencing via stonith_admin
When the cluster fences a node, it also arranges for any devices it has
to be moved to an alternate host before executing it.

However stonith_admin has no such capabilities and if a single device
that can fence all nodes is used, and that device is only registered on
the target, then fencing will fail unless we allow nodes to shoot
themselves.