Medium: fencing: Introducing the concept of verified stonith devices.
The concept of verified stonith devices is an internal flag used
to signify what devices have successfully been contacted by a node,
and what devices have not.  When a remote operation queries the peers
for all the devices capable of fencing a specific target, the devices
that are confirmed as being verified by the peer are given priority
over un-verified devices.

Now that stonith devices configured in the CIB exist on all stonith
instances, the 'verified' concept lets remote operations give priority
to the devices on nodes pacemaker has "started" (started just means
pacemaker has run the monitor operation successfully).