Low: cts: Disable NearQuorumPoint when testing CMAN
If the cluster gains quorum but does not have every node active, and
immediately afterwards the node with the lowest nodeid attempts to shut
down...

- /Pacemaker/ on node 1 is busy shutting down while /cman/ is coming
  up on nodes 3 and 4.
- The overlap allows the cluster to obtain quorum.
- Once quorum is obtained, dlm/fenced (on node 1 because it has the
  lowest nodeid) decides it needs to shoot node 2
- fenced invokes fence_pcmk which relays the request to stonithd

This will fail because stonithd isn't running and fenced will retry
forever, blocking shutdown.

This problem is specific to CMAN clusters with more than two nodes.