High: Remove functionality for terminating old lrmd and pengine instances
If corosync or heartbeat dies, anything with a connection to it will
exit.  The pengine and lrmd are not cluster aware and so would hang
around.

Previously the new instance would arrange for the old to terminate,
however there is little point as the node would normally be fenced.

So don't bother anymore, especially as respawning all the daemons
creates additional downtime when pacemakerd is involved.

(A separate patch allows pacemakerd to 're-attach' to existing daemons
at startiup)