High: crmd: Prevent election storms caused by getrusage() values being too close
If nodes are started at exactly the same time, their getrusage() values
could be too close, such that the "cost" of processing a peers vote
causes the local value to now exceed that of the peer.  If the same
thing happens on the peer, both sides will not be able to agree on a
winner.