Medium: fencing: Don't start fencing op until a promising query result has arrived
This adds logic for broadcast fencing that is similar to what is already
done for topology mode.  For broadcast mode, query results are now
gathered until a promising result comes in or the query timeout
is triggered. This helps prevent executing node suicide when better
query results might be available.