Fix: crmd: Prevent use-after-free when the blackbox is enabled
A field from xmlNode was being passed to the callsite as 'function'.
At the point the xmlNode was free'd the cs->function pointer was no
longer valid.  Logging from the same point would then result in a
comparision with the dangling pointer.

Fixed in
https://github.com/asalkeld/libqb/commit/1c9104e334fcf9ee252632e1801c701d60e3c354
but there was no real advantage to not using __FUNCTION__ instead.