Fix: cib: Strip text nodes on writing CIB to disk
The problem happens if there are any text nodes in the CIB XML.

In write_cib_contents(), after writing the CIB and the digest to disk,
it'll instantly re-check them via retrieveCib(). While the text nodes
will be stripped when reading from the on disk XML file, which will
result in a mis-matched digest. So the write fails, and cib disk write
will be disabled.

This commit strips text nodes on writing CIB to disk to ensure the
consistency.