Fix: crmd: Abort transition if the resource of a returned action no longer exists in the lrmd
If a cleanup is performed while a start action is in-flight, when the
start action returns, it won't do cib update since the lrm resource no
longer exists. So tengine won't be aware of the completion of the
action. It will get stuck in S_POLICY_ENGINE state until the action
timer pops. That's not acceptable if the start timeout is quite long.