Low: remote: Properly handle expected late replies when TLS transport is in use.
Because the request/response protocol in libqb requires a response for
every request, the TLS transport needs to work the same way. For the
few situations where we don't actually care what the response is, we
should be able to send the request over TLS and not wait around for the
response.  This patch allows the TLS client to do this without getting
warnings about late responses later on.