High: pengine: Allow fencing of baremetal remote nodes
Baremetal nodes are now fenced in the same way cluster nodes
are fenced.  If the baremetal node has not status section, it
will be fenced on startup. If the baremetal node is down but
there are resources that didn't shutdown, the node will be fenced.
If a resource failure on the baremetal node should result in
the node being fenced on failure, then the baremetal node will
get fenced.