Feature: crmd: Support "actions-limit" - the number of jobs that the TE is allowed to execute in parallel on a node
Except "batch-limit", which is a cluster-wide concurrency limit,
there are two per-node concurrency limits can be configured now:

* actions-limit:
The number of jobs that the TE is allowed to execute in parallel on a
node. If set to zero or unset, defaults to at least 4 or twice the
number of CPU cores on the DC. -1 means unlimited.

* migration-limit:
The number of migration (migrate_to/migrate_from) jobs that the TE is
allowed to execute in parallel on a node. If set to zero or unset,
defaults to half the number of CPU cores on the DC. -1 means unlimited.

1. They can be configured as a global property in crm_config.
2. They can be configured as a per-node attribute, which can override
the global property.
3. In practical, also depending on the configuration, any of them can
be reached first.