Fix: crmd: kill pengine on shutdown also when using heartbeat
If this is a heartbeat cluster, kill the pengine
in pe_ipc_destroy already.

The pe_subsystem->pid was set to -1 too early,
in pe_ipc_destroy, but without actually killing it.

Now stop_subsystem() in do_stop() would find it "already stopped",
and do nothing.

We would "leak" a pengine process, which may lead to problems
on next cluster start, especially when there was an update meanwhile.
The still running pengine may use outdated libraries/binaries.

There may be other problems as well when re-attaching
to a still running pengine.