Fix: Probe for baremetal remote-node connection resources.
Baremetal remote-node connection resources can live anywhere
in the cluster. Unlike the connection resources associated with
container nodes, the location of baremetal connection resources
by default are not predictable.  This means if the lrm status involved
with a baremetal connection resource is ever cleared (crm_resource -C)
the active baremetal connection will persist, yet the cluster will have
no knowledge of it.

To avoid this, probes have been enabled for baremetal connection
resources so those connections that get cleared in the cib can be
rediscoved if they were not stopped in the crmd.