High: pengine: Recover unexpectedly running container nodes.
Container node probes have been disabled because probing into
containers has proven to introduce transition conflicts when
certain ordering constraints are in use.  Probing into container
nodes is not entirely necessary though if we can be certain that
when the container node starts, it is clean of any cluster resources.

This patch allows us to safely assume that container nodes will
always be started and controlled by the cluster.  If a container node
is started outside of the control of the cluster, the cluster will
recover the container node to regain and guarantee control over it.

In the future, we would like to enable probing into container nodes
again.  There a substantial amount of work to be done involving
the order in which probes are executed before container probes can
become a reality.