crm_attribute: Default --lifetime to forever when using --node
None of the examples using --node in the help text will succeed
since read_attr_delegate() is passed NULL as section and fails
on an assertion. section should be set using --lifetime to either
"reboot" or "forever". This patch makes --lifetime default to
"forever" if the value was not set explicitly.