Fix: remote: Handle request/response ipc proxy correctly
We now preserve whether a proxied ipc server message is intended
for the event or response channel without trying to guess based
upon the way we expect most server implementations to behave.

This fixes an issue where pacemaker_remote is unable to manage
master/slave resources because proxy communication with the attrd
is unreliable.