High: remote: clear remote-node status correctly
Previously container remote-nodes had their status cleared
after the connection resource started. If the transition is aborted
between the container starting and the connection resource starting
for a remote-node, the previous status entry causes it to look like
the container is up with active resources while the connection resource
is down... This causes the container to be fenced.

To avoid this situation, container remote-nodes initialize their
status sections right as the container is up, even if the actual
connection resource has not started.