High: lrmd: pass exit reason prefix to ocf scripts as env variable
Previously we did a bunch of build dependency logic to determine
what the exit reason prefix string is by pulling it out of the
resource-agents package. Now the resource-agents package has
been updated to take in a dynamic exit reason prefix as an
environment variable.  By using this method, we avoid having to
maintain the package dependency.