High: pengine: relaxed-probes metadata option
The relaxed-probes resource metadata option reduces the number
of nodes a resource needs to be probed on by only probing nodes
that the resource has a >= 0 location score on. The assumption
is that a resource with this option set will be physically
incapable of running on a node with a negative location constraint
score.