fix crm_update_peer_proc to NOT ignore flags if partially set
The "set_bit()" function used here actually deals with masks, not bit numbers.
The "flag" argument should in fact be plural: flags.

These proc flag bits are not always set one at a time,
but for example as "crm_proc_crmd | crm_proc_cpg",
and not necessarily cleared with the same combination.

Ignoring to-be-set flags just because *some* of the flag bits are
already set is clearly a bug, and may be the reason for stale process
cache information.