cib: refactor: separate daemon-specific code from generic code
This revision refactors write_cib_contents() to not use crm_rename()
or retrieveCib() (along with misc. simplifications), simplifies
those two functions to the daemon-specific use case, and removes
validate_on_disk_cib().

The goal is to avoid using the daemon-specific global variable
cib_writes_enabled inside the non-daemon-specific portion of
write_cib_contents(), which is planned to be moved to the CIB library.