tools: display node names more consistently in crm_mon output
This refactors a new function get_node_display_name() to display a node name
(with node ID and/or container ID when appropriate) and modifies the
plaintext and HTML outputs to use it when displaying node names. Also,
the XML output will now add a container_id attribute to node tags
when appropriate.