High: pengine: pass node name of pcmk host into isolated resources
This feature passes the hostname of the pcmk node that is executing
an isolated resource action into the isolated resource's environment.
The hostname is passed in as CRM_meta_isolation_host environment
variable. When things like crm_node or crm_attribute encounter
this environment variable, they know to use the value in that varible
to represent the pcmk node name. This allows isolated resources to set
transient attributes.