Low: pengine: update regression tests to account for promote restart fix
master-5
master-6
Both of these tests seem to be affected by a bug where the digest was
not present in the promote operation. Since the digest didnt' mean
anything for promote until now, this went unnoticed. I manually entered
the digest for these two tests so they would perform as intended.

bug-1572-2
This regression test failed for a similar reason. The digest was recorded
and actually did change. A restart should have occurred, but since the
digest was never comparied for the promote, it never happend. Since this
restart is not what the test is intended to verify, I manually made the
digests work so the test will export the same results