Refactor: crmd: remove dead code
Don't check for op_allow_fail in fencing operations (a usage that was
removed long ago in 0a394a8 and fb6fb35) so we can repurpose op_allow_fail
later for better handling of on-fail=ignore. While we're at it,
remove an unused argument from match_graph_event(), and add more comments.