Fix: crmd: Prevent the old version of DC from being fenced when it shuts down for rolling-upgrade
This issue is related to 5cb541f and 23b5f07.

In a mixed cluster, if the node that is shutting down is an old version
of DC, it won't send the CRM_OP_SHUTDOWN_REQ to the other newly upgraded
nodes. So the other nodes won't get a chance to update this node's
"expected" field to "down" from their membership caches, while they'll
still erase this node's transient attributes. Any node that is later
elected to be the new DC will decide to fence this node that has just
shut down.

The idea of this fix is not to erase this node's transient attributes
in this case to prevent the fencing. Although when this node first
rejoins, there could be the potential race condition. That's still
better than being fenced I think.