Fix: cib: Attempt to disable legacy mode for plugin-based cluster after a node has left the cluster
The idea was introduced by 4f00ccc. But it seems to only work for cpg
enabled cluster. For a plugin-based cluster, after rolling-upgrade, the
nodes that used to be in the mixed cluster are still in the cib legacy
mode, while the last upgraded node is in the new cib mode. Then the cib
can only get synchronized when the node issues a cib ping operation.