Fix: PE: Skip unrunnable actions when one-or-more is in effect
The require-all=false/one-or-more logic will already mark the psuedo op
as unrunnable if the one-or-more criteria cannot be satisfied.

Therefor we can simply exclude these unrunnable actions from the graph.

They appear whenever clone-max is greater than the number of instances
that will ever be run (ie. due to location constraints)