Build: prefer more effective Docbook template discovery
On systems equipped with xmlcatalog binary (standard part of libxml2
that should be a dependency for libxml2-devel already required at
build-time, anyway), it seems optimal to query a namespace URI dedicated
to respective Docbook XSL templates (using "current" pointer rather than
anything static) and then using a local "mirror" path for a direct access.

If such an attempt fails, original brute-force method is used, with
non-determinism[*] partly mitigated in a way that a list of candidates
for a match is sorted prior to its usage for iteration.

[*] non-determinism may, e.g., arise in Fedora with both
    docbook-style-xsl and docbook5-style-xsl packages installed;
    admittedly the only difference in generated man pages is then:

    -.\" Generator: DocBook XSL-NS Stylesheets v1.78.1 <http://docbook.sf.net/>
    +.\" Generator: DocBook XSL Stylesheets v1.78.1 <http://docbook.sf.net/>

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>