Refactor: fencing: make structure member name for device count clearer
This renames the "devices" member to "ndevices" in st_query_result_t.
It's a little easier to read since "devices" is often used in other contexts as
a list or such, but the main reason is because it allows the current
"device_list" member be renamed to "devices" if its implementation changes
as planned for remapping reboot to off+on.