Refactor: fencing: mark devices as done rather than remove them
Previously, once a fencing device was executed, it was removed from the
peer result's list of available devices. Now, the device properties
includes an "executed" member that is set to true instead.

This has no effect now, but will be required when remapping reboot
to off+on, because the device list will need to be reused in the
"on" phase.