Refactor: fencing: pass operation rather than default timeout to timeout functions
Previously, the various fencing timeout functions took a default_timeout argument
that was always set (directly or indirectly) to op->base_timeout. Now, op is
passed instead, and the functions use op->base_timeout directly.

This has no effect now, but is necessary for remapping reboot to off+on later,
because other operation information (whether we're in the off or on phase)
will be needed to calculate the right timeout.