Feature: fencing: add support for operation phases
This adds support for having multiple "phases" of an operation in the
remote_fencing_op_t structure. This will be used to support remapping
sequential topology reboots to all-off-then-all-on.

This is not complete support, only support within the one structure and
associated code, so at this point, phase is always 0 and these changes have
no effect.